refactored some code, started typography components
