+++ frontend/src/modules/Components.tsx
@@ -10,6 +10,7 @@ import {
   Radio,
   MenuItem,
   SelectChangeEvent,
+  Typography,
 } from '@mui/material'
 import { useState } from 'react'
 import SearchBar from '@core/Components/SearchBar'
@@ -175,6 +176,18 @@ const Components = () => {
       </Box>
       <Box my={3}>
         <ProgressBar total={8} step={4} />
+      </Box>
+      <Box my={3}>
+        <Typography variant="h1">Heading 1</Typography>
+        <Typography variant="h2">Heading 2</Typography>
+        <Typography variant="h3">Heading 3</Typography>
+        <Typography variant="h4">Heading 4</Typography>
+        <Typography variant="h5">Heading 5</Typography>
+        <Typography variant="h6">Heading 6</Typography>
+        <Typography variant="subtitle1">Subtitle 1</Typography>
+        <Typography variant="subtitle2">Subtitle 2</Typography>
+        <Typography variant="body1">Body text 1</Typography>
+        <Typography variant="body2">Body text 2</Typography>
       </Box>
     </Box>
   )
